requirements.txt
streamlit
Pillow
transformers
pytesseract
opencv-python-headless
numpy
gTTS
torch